remove some global functions
Committed-by: xiaolei.zl from Dev container

Committed-by: xiaolei.zl from Dev container
