+++ src/utils/property/column.cc
@@ -17,7 +17,7 @@
 
 #include <limits>
 
-#include "neug/storages/container_utils.h"
+#include "neug/storages/container/container_utils.h"
 #include "neug/utils/id_indexer.h"
 #include "neug/utils/property/table.h"
 #include "neug/utils/property/types.h"
